remove visviva.py as useless
